+++ src/main/resources/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Ancient Trees Changelog
 
+## 1.5.1
+- Fixed crash when Storage Drawers integration is enabled and Refined Relocation is present.
+
+## 1.5.0
+- Storage Drawers Integration: Make storage drawers from ancient trees!
+- Switches in configuration file to toggle integration with other mods.
+
 ## 1.4.0
 - Saplings can now be planted, fertilized and harvested in MineFactory reloaded machines. (Though you should
   be careful with larger trees!)